testing_payment/requirements.txt
Flask==3.0.0
stripe==11.2.0
python-dotenv==1.0.1
